Change 'nilai_skp' column type from tinyInteger to decimal in daftar_penilaian_rewards table

+++ database/migrations/2024_11_24_202008_create_daftar_penilaian_rewards_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('daftar_penilaian_rewards', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('nilai_skp')->nullable()->unsigned();
+            $table->decimal('nilai_skp')->nullable()->unsigned();
             $table->tinyInteger('tk')->nullable()->unsigned();
             $table->tinyInteger('tl1')->nullable()->unsigned();
             $table->tinyInteger('tl2')->nullable()->unsigned();
